+++ core/geometry/src/curve/GeometryQuery.ts
@@ -16,7 +16,8 @@ import { CurveCollection } from "./CurveCollection";
 import { CurvePrimitive } from "./CurvePrimitive";
 import { PointString3d } from "./PointString3d";
 
-/** Describes the category of a [[GeometryQuery]], enabling type-switching like:
+/**
+ * Describes the category of a [[GeometryQuery]], enabling type-switching like:
  * ```ts
  *   function processGeometryQuery(q: GeometryQuery): void {
  *     if ("solid" === q.geometryCategory)
@@ -38,56 +39,54 @@ import { PointString3d } from "./PointString3d";
  */
 export type GeometryQueryCategory = "polyface" | "curvePrimitive" | "curveCollection" | "solid" | "point" | "pointCollection" | "bsurf";
 
-/** Union type for subclasses of [[GeometryQuery]]. Specific subclasses can be discriminated at compile- or run-time using [[GeometryQuery.geometryCategory]].
+/**
+ * Union type for subclasses of [[GeometryQuery]]. Specific subclasses can be discriminated at compile- or run-time
+ * using [[GeometryQuery.geometryCategory]].
  * @public
  */
 export type AnyGeometryQuery = Polyface | CurvePrimitive | CurveCollection | SolidPrimitive | CoordinateXYZ | PointString3d | BSpline2dNd;
 
-/** Queries to be supported by Curve, Surface, and Solid objects */
 /**
- * * `GeometryQuery` is an abstract base class with (abstract) methods for querying curve, solid primitive, mesh, and bspline surfaces
+ * Queries to be supported by Curve, Surface, and Solid objects
+ * * `GeometryQuery` is an abstract base class with (abstract) methods for querying curve, solid primitive, mesh,
+ * and bspline surfaces
  * @public
  */
 export abstract class GeometryQuery {
   /** Type discriminator. */
   public abstract readonly geometryCategory: GeometryQueryCategory;
-
-  /** return the range of the entire (tree) GeometryQuery */
+  /** Return the range of the entire (tree) GeometryQuery */
   public range(transform?: Transform, result?: Range3d): Range3d {
     if (result) result.setNull();
     const range = result ? result : Range3d.createNull();
     this.extendRange(range, transform);
     return range;
   }
-
-  /** extend rangeToExtend by the range of this geometry multiplied by the transform */
+  /** Extend rangeToExtend by the range of this geometry multiplied by the transform */
   public abstract extendRange(rangeToExtend: Range3d, transform?: Transform): void;
-
-  /** Attempt to transform in place.
-   *
+  /**
+   * Attempt to transform in place.
    * * LineSegment3d, Arc3d, LineString3d, BsplineCurve3d always succeed.
    * * Some geometry types may fail if scaling is non-uniform.
    */
   public abstract tryTransformInPlace(transform: Transform): boolean;
-
-  /** try to move the geometry by dx,dy,dz */
+  /** Try to move the geometry by dx,dy,dz */
   public tryTranslateInPlace(dx: number, dy: number = 0.0, dz: number = 0.0): boolean {
     return this.tryTransformInPlace(Transform.createTranslationXYZ(dx, dy, dz));
   }
-  /** return a transformed clone.
-   */
+  /** Return a transformed clone. */
   public abstract cloneTransformed(transform: Transform): GeometryQuery | undefined;
-  /** return a clone */
+  /** Return a clone */
   public abstract clone(): GeometryQuery | undefined;
-  /** return GeometryQuery children for recursive queries.
-   *
+  /**
+   * Return GeometryQuery children for recursive queries.
    * * leaf classes do not need to implement.
    */
   public get children(): GeometryQuery[] | undefined { return undefined; }
-  /** test if (other instanceof this.Type).  REQUIRED IN ALL CONCRETE CLASSES */
+  /** Test if (other instanceof this.Type).  REQUIRED IN ALL CONCRETE CLASSES */
   public abstract isSameGeometryClass(other: GeometryQuery): boolean;
-  /** test for exact structure and nearly identical geometry.
-   *
+  /**
+   * Test for exact structure and nearly identical geometry.
    * *  Leaf classes must implement !!!
    * *  base class implementation recurses through children.
    * *  base implementation is complete for classes with children and no properties.
@@ -113,7 +112,8 @@ export abstract class GeometryQuery {
     }
     return false;
   }
-  /** apply instance method isAlmostEqual if both are defined.
+  /**
+   * Apply instance method isAlmostEqual if both are defined.
    * * both undefined returns true
    * * single defined returns false
    */
@@ -126,8 +126,10 @@ export abstract class GeometryQuery {
   }
   /**
    * * "double dispatch" call pattern.
-   * * User code implements a `GeometryHandler` with specialized methods to handle `LineSegment3d`, `Arc3d` etc as relevant to its use case.
-   * * Each such `GeometryQuery` class implements this method as a one-line method containing the appropriate call such as `handler.handleLineSegment3d ()`
+   * * User code implements a `GeometryHandler` with specialized methods to handle `LineSegment3d`, `Arc3d` etc as
+   * relevant to its use case.
+   * * Each such `GeometryQuery` class implements this method as a one-line method containing the appropriate call
+   * such as `handler.handleLineSegment3d ()`
    * * This allows each type-specific method to be called without a switch or `instanceof` test.
    * @param handler handler to be called by the particular geometry class
    */