+++ core/ecschema-editing/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log - @itwin/ecschema-editing
 
-This log was last generated on Tue, 25 Apr 2023 17:50:35 GMT and should not be manually modified.
+This log was last generated on Mon, 22 May 2023 15:34:14 GMT and should not be manually modified.
+
+## 4.0.0
+Mon, 22 May 2023 15:34:14 GMT
+
+### Updates
+
+- Updated to work with async schema loading changes made to ecschema-metadata
+- Update to eslint@8
+- Removed obsolete EC rule 102 - Abstract class must have abstract base class.
+
+## 3.7.6
+Mon, 15 May 2023 18:23:41 GMT
+
+_Version update only_
+
+## 3.7.5
+Thu, 04 May 2023 19:43:18 GMT
+
+### Updates
+
+- Removed obsolete EC rule 102 - Abstract class must have abstract base class.
 
 ## 3.7.4
 Tue, 25 Apr 2023 17:50:35 GMT